requirements.txt
fritzconnection
mysql-connector-python
speedtest-cli
requests